feat: Add post-processing script and fix torrent completion logic
Add a parameterized script to be executed on torrent completion to copy files from a local mount to a remote storage using rclone.

Fix the logic for torrent completion to allow the script to run even when no files are downloaded to the host.

Add support for %RD_ID argument to pass the Real-Debrid torrent ID to the script (and subsequently fix the argument parsing).

Ensure files are selected on Real-Debrid before running the script to generate download links.

Add curl and jq to the Dockerfile as dependencies for the script.

Update the version in package.json to 2.0.119.

Include minor bug fixes and refactoring in WatchFolderChecker and TorrentRunner.